+++ docs/CHANGE_LOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.8.2:
 - Added export support for the global "$observe" object, which can now be imported as "observe.js".
+- BREAKING: Renamed the binding attribute "data-observe-watch-options" to "data-observe-js".
 
 <br>